Big refactor - part 3 (end)
