+++ ServiceLayer/AuthorizeSetup/AuthCookieValidateRefreshClaims.cs
@@ -14,7 +14,7 @@ using FeatureAuthorize;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace ServiceLayer.AuthoriseSetup
+namespace ServiceLayer.AuthorizeSetup
 {
     /// <summary>
     /// This version provides:
